Move ElasticsearchIndexRebuilder / ElasticsearchAtomicIndexRebuilder into elasticsearch7 module
